Fixed hashing for materials and textures
